TESTS: add tests for IPv6 support
